+++ packages/jest-runner/src/runTest.ts
@@ -86,6 +86,7 @@ async function runTestInternal(
   const docblockPragmas = docblock.parse(docblock.extract(testSource));
   const customEnvironment = docblockPragmas['jest-environment'];
 
+  const loadTestEnvironmentStart = Date.now();
   let testEnvironment = projectConfig.testEnvironment;
 
   if (customEnvironment) {
@@ -168,6 +169,7 @@ async function runTestInternal(
       testPath: path,
     },
   );
+  const loadTestEnvironmentEnd = Date.now();
 
   if (typeof environment.getVmContext !== 'function') {
     console.error(
@@ -212,6 +214,7 @@ async function runTestInternal(
 
   const start = Date.now();
 
+  const setupFilesStart = Date.now();
   for (const path of projectConfig.setupFiles) {
     const esm = runtime.unstable_shouldLoadAsEsm(path);
 
@@ -224,6 +227,7 @@ async function runTestInternal(
       }
     }
   }
+  const setupFilesEnd = Date.now();
 
   const sourcemapOptions: sourcemapSupport.Options = {
     environment: 'node',
@@ -328,8 +332,13 @@ async function runTestInternal(
     const end = Date.now();
     const testRuntime = end - start;
     result.perfStats = {
+      ...result.perfStats,
       end,
+      loadTestEnvironmentEnd,
+      loadTestEnvironmentStart,
       runtime: testRuntime,
+      setupFilesEnd,
+      setupFilesStart,
       slow: testRuntime / 1000 > projectConfig.slowTestThreshold,
       start,
     };